Rename: Change react query mutation file name
